Update Vimux API usage to new pubic methods
The _VimuxOption method was a pseudo private API that should never have
been used — or rather in this case should never have been private. It is
now promoted to a regular public method. Additionally defaults are
handled in such a way that they don't need to be specified on every
possible use of an option.